% Generated by roxygen2 (4.0.1): do not edit by hand
\name{inlineEditor}
\alias{inlineEditor}
\title{Inline Code Editor Widget}
\usage{
inlineEditor(type = c("plot", "text"), codeWidth = 8, initialVal = "",
  height = 300)
}
\arguments{
\item{type}{Either \code{'plot'} or \code{'text'}, depending on what you want
to output.}

\item{codeWidth}{Out of 12, how many columns should be occupied by the code,
the rest will be used for the output.}

\item{initialVal}{The initial text to put in the editor.}

\item{height}{The height of the editor (in pixels).}
}
\description{
Creates an inline code editor using the \pkg{shinyAce} package that allows
users to dynamically enter and evaluate R code and see either a plot or
textual output.
}

